pyproject.toml
[project]
name = "telegrambot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.1",
    "flake8>=7.3.0",
    "pprintpp>=0.4.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
    "requests>=2.32.5",
    "rich>=14.3.3",
]
